Refactor ticket price datatype to BigDecimal for precision

+++ src/main/java/de/croebe/tickets/domain/dtos/ListTicketTicketTypeResponseDto.java
@@ -1,10 +1,11 @@
 package de.croebe.tickets.domain.dtos;
 
+import java.math.BigDecimal;
 import java.util.UUID;
 
 public record ListTicketTicketTypeResponseDto(
         UUID id,
         String name,
-        Double price
+        BigDecimal price
 ) {
 }
